% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{baWmean}
\alias{baWmean}
\title{baWmean}
\usage{
baWmean(modOut, varX)
}
\arguments{
\item{modOut}{PREBAS output of a multisite run}

\item{varX}{index of the variable for which the basal area weighted mean needs to be calculated}
}
\value{
The weighted mean basal area for a PREBAS output (including both multi site and single site runs)
}
\description{
baWmean
}
