% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{calBAspec}
\alias{calBAspec}
\title{calBAspec function}
\usage{
calBAspec(prebout, speciesID)
}
\arguments{
\item{SpecieID}{ID number of the layer for the specie}

\item{Prebasoutput}{route or variable to the data needed}
}
\value{
the Basal area for the Specie given
}
\description{
calBAspec function
}
