% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeBioInd.R
\name{calBioIndices}
\alias{calBioIndices}
\title{calBioIndices function}
\usage{
calBioIndices(modOut)
}
\arguments{
\item{modOut}{The Prebass output}
}
\value{
the habitat suitability indices
}
\description{
calBioIndices function
}
