% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{calVdeadSpec}
\alias{calVdeadSpec}
\title{calVdeadSpec function}
\usage{
calVdeadSpec(prebout, speciesID)
}
\arguments{
\item{SpecieID}{ID number of the layer for the specie}

\item{Prebasoutput}{route or variable to the data needed}
}
\value{
the volume for the specie given
}
\description{
calVdeadSpec function
}
