% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mai.r
\name{compMAI}
\alias{compMAI}
\title{compMAI}
\usage{
compMAI(modRun)
}
\arguments{
\item{modRun}{multiSite PREBAS output}

\item{#'}{}
}
\value{
mean annual increment of the stand (standMAI) and bylayer (layerMAI)
}
\description{
compMAI
}
\examples{
prebOut <- TransectRun(species="Mixed"); compMAI(prebOut)
}
