% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{DqFunMS}
\alias{DqFunMS}
\title{DqFunMS}
\usage{
DqFunMS(modOut)
}
\arguments{
\item{modOut}{PREBAS output of a multisite run}
}
\value{
The quadratic mean diameter of a multisite PREBAS run
}
\description{
DqFunMS
}
