% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestNavigatorManagementFunctions.R
\name{forest_management_update}
\alias{forest_management_update}
\title{management function updater (ForestNavigator)}
\usage{
forest_management_update(
  initPrebas,
  forest_type_management_tab,
  country,
  management,
  est_bau_pars = est_bau_pars_def
)
}
\arguments{
\item{initPrebas}{Rprebasso initialization object for multisite created by the InitMultiSite function}

\item{forest_type_management_tab}{Table with forest management (for_man) by forest type (forest_type) and siteID (site)}

\item{country}{Country of simulations. Choose between: Sweden}

\item{management}{type of management. choose between: BAU}
}
\description{
management function updater (ForestNavigator)
}
