% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{HSIcaper}
\alias{HSIcaper}
\title{Habitat suitability for the Capercaillie}
\usage{
HSIcaper(Vpine, Vspruce, Nstems)
}
\arguments{
\item{Vpine}{volume (m3/ha) of Pine}

\item{Vspruce}{volume (m3/ha) of Spruce}

\item{Nstems}{total number of trees per hectare}
}
\value{
the habitat suitability index
}
\description{
Habitat suitability for the Capercaillie
}
