% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{HSIfs}
\alias{HSIfs}
\title{Habitat suitability for the Siberian flying squirrel}
\usage{
HSIfs(Vspruce, Pspruce, Vdec)
}
\arguments{
\item{Vspruce}{volume (m3/ha) of Spruce}

\item{Pspruce}{Proportion of spruce (percentage) of the total tree volume}

\item{Vdec}{volume (m3/ha) of deciduous trees}
}
\value{
the habitat suitability index
}
\description{
Habitat suitability for the Siberian flying squirrel
}
