% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{HSIhg}
\alias{HSIhg}
\title{Habitat suitability for the Hazel grouse}
\usage{
HSIhg(ageoldest, Pdec, Pspruce)
}
\arguments{
\item{ageoldest}{Age of the oldest layer}

\item{Pdec}{Proportion of deciduous tress (percentage) of the total tree volume}

\item{Pspruce}{Proportion of spruce (percentage) of the total tree volume}
}
\value{
the habitat suitability index
}
\description{
Habitat suitability for the Hazel grouse
}
