% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{HSIlswo}
\alias{HSIlswo}
\title{Habitat suitability for the Lesser-spotted woodpecker}
\usage{
HSIlswo(BArecddec, ageoldest)
}
\arguments{
\item{BArecddec}{Total basal area (m2/ha) of recently died deciduous trees}

\item{ageoldest}{Age of the oldest layer}
}
\value{
the habitat suitability index
}
\description{
Habitat suitability for the Lesser-spotted woodpecker
}
