% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{HSIlswoPop}
\alias{HSIlswoPop}
\title{Habitat suitability for the Lesser-spotted woodpecker simplified}
\usage{
HSIlswoPop(BArecddec)
}
\arguments{
\item{BArecddec}{Total basal area (m2/ha) of recently died deciduous trees}
}
\value{
the probability of presence of the bird
}
\description{
Habitat suitability for the Lesser-spotted woodpecker simplified
}
