% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{HSIltt}
\alias{HSIltt}
\title{Habitat suitability for the Long-tailed tit}
\usage{
HSIltt(ageoldest, BAtot, Pdec)
}
\arguments{
\item{ageoldest}{Age of the oldest layer}

\item{BAtot}{Total basal area (m2/ha)}

\item{Pdec}{Proportion of deciduous tress (percentage) of the total tree volume}
}
\value{
the habitat suitability index
}
\description{
Habitat suitability for the Long-tailed tit
}
