% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{HSIttwo}
\alias{HSIttwo}
\title{Habitat suitability for the Three-toed woodpecker}
\usage{
HSIttwo(BArecd, Vtotal)
}
\arguments{
\item{BArecd}{Total basal area (m2/ha) of recently died trees}

\item{Vtotal}{Total timber volume of living trees}
}
\value{
the habitat suitability index
}
\description{
Habitat suitability for the Three-toed woodpecker
}
