% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioInd.r
\name{HSIttwoPop}
\alias{HSIttwoPop}
\title{Habitat suitability for the Three-toed woodpecker simplified}
\usage{
HSIttwoPop(BArecd)
}
\arguments{
\item{BArecd}{Total basal area (m2/ha) of recently died trees}
}
\value{
the probability of presence of the bird
}
\description{
Habitat suitability for the Three-toed woodpecker simplified
}
