% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{initialAgeSeedl}
\alias{initialAgeSeedl}
\title{Initial age of the seedlings}
\usage{
initialAgeSeedl(SiteType, ETS)
}
\arguments{
\item{SiteType}{Site type}

\item{ETS}{Effective temperature sum}
}
\value{
Initial age of the seedlings
}
\description{
Initial age of the seedlings
}
\examples{
initialAgeSeedl(2, 1200)
}
