% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{initPrebas_subsetter}
\alias{initPrebas_subsetter}
\title{Title}
\usage{
initPrebas_subsetter(initPrebas, sitex)
}
\arguments{
\item{initPrebas}{multisite Initialization object}

\item{sitex}{sites to be selected}
}
\value{
a multisite prebas initialization object with just the sitex sites
}
\description{
Title
}
