% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yasso_routines.r
\name{initSoilC_fromTot}
\alias{initSoilC_fromTot}
\title{initSoilC_fromTot
this function splits the soilC in the AWENH pools used in YASSO. 
the output can then be used in PREBAS runs as initial soilC}
\usage{
initSoilC_fromTot(
  soilCTot,
  organShares = p_organShares,
  awenhShares = p_awenhShares
)
}
\arguments{
\item{soilCTot}{}

\item{organShares}{}

\item{awenhShares}{}
}
\description{
initSoilC_fromTot
this function splits the soilC in the AWENH pools used in YASSO. 
the output can then be used in PREBAS runs as initial soilC
}
