% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{mLit}
\alias{mLit}
\title{mLit}
\usage{
mLit(aLit, months = 8:9)
}
\arguments{
\item{aLit}{Number of years}

\item{months}{August and September}
}
\value{
A time series with the annual autumn litter fall.
}
\description{
mLit
}
