% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{monthlyFluxes}
\alias{monthlyFluxes}
\title{monthlyFluxes}
\usage{
monthlyFluxes(modOut, weatherOption = 1)
}
\arguments{
\item{modOut}{PREBAS output of a multisite run}

\item{weatherOption}{}
}
\value{

}
\description{
monthlyFluxes
}
