% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{monthlyGPP}
\alias{monthlyGPP}
\title{monthlyGPP}
\usage{
monthlyGPP(GPPx, func = "sum")
}
\arguments{
\item{GPPx}{}

\item{func}{}
}
\value{
A vector of the mean daily GPP for each month.
}
\description{
monthlyGPP
}
