% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{monthlyWeather}
\alias{monthlyWeather}
\title{monthlyWeather}
\usage{
monthlyWeather(varx, func)
}
\arguments{
\item{varx}{}

\item{func}{}
}
\value{
A vector of the mean daily weather for each month.
}
\description{
monthlyWeather
}
