% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{NesterovInd}
\alias{NesterovInd}
\title{Nesterov Index
A cumulative function of daily Tmax and dew-point temperature Tdew, eq. 5 in TH2010}
\usage{
NesterovInd(rain, tmin, tmax)
}
\arguments{
\item{rain}{daily precipitation (mm)}

\item{tmin}{daily minimum temperature (ºC)}

\item{tmaX}{daily maximum temperature (ºC)}
}
\value{
the Nesterov Index
}
\description{
Nesterov Index
A cumulative function of daily Tmax and dew-point temperature Tdew, eq. 5 in TH2010
}
