% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSitePrebas.r
\name{regionPrebas}
\alias{regionPrebas}
\title{Title}
\usage{
regionPrebas(
  multiSiteInit,
  HarvLim = NA,
  minDharv = 999,
  cutAreas = NA,
  compHarv = 0,
  thinFact = 0.25,
  ageHarvPrior = 0,
  siteOrder = NA,
  fertThin = 0,
  nYearsFert = 20,
  yearFert = NULL,
  deltaSiteTypeFert = 1,
  oldLayer = 0,
  startSimYear = 1,
  frac_clct = NULL
)
}
\arguments{
\item{frac_clct}{fraction of clearcutted area over the total area of the region}
}
\description{
Title
}
