% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{reineke1}
\alias{reineke1}
\title{reineke1}
\usage{
reineke1(Dq, k, spID)
}
\arguments{
\item{Dq}{quadratic mean diameter of a multisite PREBAS run}

\item{k}{}

\item{spID}{}
}
\value{
Reineke Stand density index
}
\description{
reineke1
}
