% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSitePrebas.r
\name{reStartRegionPrebas}
\alias{reStartRegionPrebas}
\title{Title}
\usage{
reStartRegionPrebas(
  multiSiteInit,
  HarvLim = NA,
  minDharv = 999,
  cutAreas = NA,
  compHarv = 0,
  thinFact = 0.25,
  ageHarvPrior = 0,
  siteOrder = NA,
  fertThin = 0,
  yearFert = NULL,
  deltaSiteTypeFert = 1,
  nYearsFert = 20,
  oldLayer = 0,
  startSimYear
)
}
\arguments{
\item{multiSiteInit}{}

\item{HarvLim}{}

\item{minDharv}{}

\item{cutAreas}{}

\item{compHarv}{}

\item{thinFact}{}

\item{ageHarvPrior}{}

\item{siteOrder}{}

\item{fertThin}{}

\item{nYearsFert}{}

\item{oldLayer}{}

\item{startSimYear}{}
}
\description{
Title
}
