% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteType.R
\name{SiteType}
\alias{SiteType}
\title{Estimate site type based on age, height, and climate.}
\usage{
SiteType(
  Age = NA,
  Hdom = NA,
  Hmean = NA,
  Species = NA,
  PAR = NA,
  TAir = NA,
  VPD = NA,
  Precip = NA,
  CO2 = NA,
  pCROBAS = pCROB,
  Age.1.5m = 5
)
}
\arguments{
\item{Age}{Age of dominant trees or mean trees (yr).}

\item{Hdom}{Dominant height (m).}

\item{Hmean}{Average height (m). If dominant height is not given, the dominant height will be calculated based on average height.}

\item{Species}{Species, ranging from 1 to 10, according to the column of pCROB.}

\item{PAR}{A matrix of daily PPFD, nrow=5, ncol=365*years.}

\item{TAir}{A matrix of daily air temperature, nrow=5, ncol=365*years.}

\item{VPD}{A matrix of daily VPD, nrow=5, ncol=365*years.}

\item{Precip}{A matrix of daily precipitation, nrow=5, ncol=365*years.}

\item{CO2}{A matrix of daily CO2 concentration, nrow=5, ncol=365*years.}

\item{pCROBAS}{Parameters of CROBAS.}

\item{Age.1.5m}{Age of seedlings when the height reach 1.5m.}
}
\value{
Site type
}
\description{
Estimate site type based on age, height, and climate.
}
\examples{
SiteType(Age=c(50,70),Hdom=c(25,30),Species=3,
     PAR = PAR,TAir= TAir,VPD= VPD,Precip= Precip,CO2= CO2,
     pCROBAS=pCROB,Age.1.5m=5)

}
