% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTable_Pine_VV.R
\name{YieldTable_Pine_VV}
\alias{YieldTable_Pine_VV}
\title{Yield table of Scots pine}
\usage{
YieldTable_Pine_VV(H100, T0, H0, N0, G0, Tn, thin.n, thin.times, thin.ints)
}
\arguments{
\item{H100}{Bonity class (Site Index)}

\item{T0}{Initial Biological age (a)}

\item{H0}{Initial Dominant height (m)}

\item{N0}{Initial Number of stems (/ha)}

\item{G0}{Initial Basal area at breast height WITH BARK (m2/ha)}

\item{Tn}{Rotation length (biological age in the end of the rotation, a)}

\item{thin.n}{Number of thinnings; if no thinnings are performed, set thin.n <- 0 and thin.times <- NULL}

\item{thin.times}{Timing of thinnings (in years of age)}

\item{thin.ints}{Intensities of thinnings (relative amount of removal IN STEM VOLUME WITH BARK); if no thinnings are performed, set thin.ints <- NULL}
}
\value{
A dataframe of yield table for Scots pine.
\itemize{
  \item T - Biological age (a)
  \item H - Dominant height (m)
  \item N - Number of stems (/ha)
  \item G - Basal area at breast height WITH BARK (m2/ha)
  \item Gu - Basal area at breast height WITHOUT BARK (m2/ha)
  \item V - Stem volume with bark (m3/ha)
  \item Bperc - Volume without bark per volume with bark (%)
  \item vbar - Mean size of stem (dm3)
  \item S - Sawtimber volume with bark (S, m3/ha).
  \item F - Pulpwood volume with bark (FF, m3/ha; minimum top diameter 6 cm)
}
}
\description{
Computing the temporal development of stand characteristics in Scots pine with the models by Vuokila & Väliaho (1980)
}
\examples{
The default setting of 5 yield tables, site index from 33 to 21. 
The variables H100, T0, H0, N0, G0 are recommended to use these default settings.

   YieldTable_Pine_VV(H100=33,T0=15,H0=5.7,N0=2200,G0=4.9,
                      Tn=110,thin.n=3,thin.times=c(50, 65, 85),thin.ints= c(0.30, 0.30, 0.30))
   
   YieldTable_Pine_VV(H100=30,T0=20,H0=7.0,N0=2200,G0=7.4,
                      Tn=110,thin.n=3,thin.times=c(50, 65, 85),thin.ints= c(0.30, 0.30, 0.30))
   
   YieldTable_Pine_VV(H100=27,T0=25,H0=7.7,N0=2000,G0=8.4,
                      Tn=110,thin.n=3,thin.times=c(50, 65, 85),thin.ints= c(0.30, 0.30, 0.30))
   
   YieldTable_Pine_VV(H100=24,T0=25,H0=5.9,N0=2000,G0=3.1,
                      Tn=110,thin.n=3,thin.times=c(50, 65, 85),thin.ints= c(0.30, 0.30, 0.30))                                                    
   
   YieldTable_Pine_VV(H100=21,T0=30,H0=6.0,N0=1800,G0=1.9,
                      Tn=110,thin.n=3,thin.times=c(50, 65, 85),thin.ints= c(0.30, 0.30, 0.30))      


}
