% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilStaff.r
\name{initBiomasses}
\alias{initBiomasses}
\title{initBiomasses}
\usage{
initBiomasses(pCro, initVarX)
}
\arguments{
\item{pCro}{matrix of CROBAS parameters where each column correspond to a different species}

\item{initVarX}{initial state of the forests.}
}
\value{
the biomasses at initialization
}
\description{
initBiomasses
}
