% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebas.r
\name{prebas}
\alias{prebas}
\title{prebas model function}
\usage{
prebas(
  nYears,
  pCROBAS = pCROB,
  pHcMod = pHcM,
  pPRELES = pPREL,
  pYASSO = pYAS,
  pAWEN = parsAWEN,
  etmodel = 0,
  siteInfo = NA,
  thinning = NA,
  initClearcut = initSeedling.def,
  fixBAinitClarcut = 1,
  initCLcutRatio = NA,
  PAR,
  TAir,
  VPD,
  Precip,
  CO2,
  P0 = NA,
  initVar = NA,
  soilC = NA,
  weatherYasso = NA,
  litterSize = litterSizeDef,
  soilCtot = NA,
  defaultThin = 1,
  ClCut = 1,
  energyCut = 0,
  inDclct = NA,
  inAclct = NA,
  yassoRun = 0,
  smoothP0 = 1,
  smoothETS = 1,
  smoothYear = 5,
  tapioPars = pTapio,
  thdPer = 0.5,
  limPer = 0.5,
  ftTapioPar = ftTapio,
  tTapioPar = tTapio,
  GVrun = 1,
  thinInt = -999,
  fertThin = 0,
  nYearsFert = 20,
  protect = 0,
  mortMod = 1,
  ECMmod = 0,
  pECMmod = parsECMmod,
  pPRELESgv = pPREL,
  layerPRELES = 0
)
}
\arguments{
\item{nYears}{number of years of the simulations}

\item{pCROBAS}{(49 x nSpecies(11) matrix) Matrix of parameter sets of CROBAS growth model, each column corresponds to a species. Default values pCROBAS = pCROB are the parameter sets for Scots pine (Pinus sylvestris), Norway spruce (Picea abies), Silver birch (Betula pendula), European beech (Fagus sylvatica), Maritime pine (Pinus pinaster), Blue gum (Eucalyptus globulus), Black locust (Robinia pseudoacacia), Populus(in Romania), Eucalyptus grandis x Eucalyptus urophylla (in Paraguay), and Norway spruce(in Germany), Quercus Ilex. Default is pCROB, print(pCROB) to see the parameter values and names.}

\item{pHcMod}{(7 x nSpecies(11) matrix) Matrix of parameter sets Height of the crown base model.}

\item{pPRELES}{A vector of PRELES parameter. Default is the boreal forest version pPREL.}

\item{pYASSO}{A vector of YASSO parameters. Default is global pYAS}

\item{pAWEN}{(12 x nSpecies(11) matrix) Matrix of parameter sets for litterfal decomposition in YASSO pools. Default is parsAWEN}

\item{etmodel}{Evapotranspiration model for PRELES. Default etmodel = 0. Possible values -1, 0, 1, 2}

\item{siteInfo}{vector of site info SiteID, climID, siteType, SWinit (initial soil water), CWinit (initial crown water), SOGinit (initial snow on ground), Sinit (initial temperature acclimation state), soildepth, effective field capacity, permanent wilthing point. Default = c(1,1,3,160,0,0,20,413.,0.45,0.118), i.e. siteType = 3.}

\item{thinning}{A matrix with thinnig inputs. Rows correspond to a thinning event. Dimensions: #Thinnings x 10. The rows, nThin, is the number of thinnings. For the columns: column 1 is year from the start of the simulation when the thinnig occur; column 2 is siteID; column 3 layer where thinnings are carried out; column 4 to 7 are stand variables (H, D, B, Hc); column 8 parameter that indicates if the stand variables (column 4:7) are provided as fraction of the actual model outputs (value=1 means that fraction is used); column 9 is the stand density after thinning if its value is not -999.; colum 10 is Sapwood area of average tree at crown base (m2) if its value is not -999 (see examples).}

\item{initClearcut}{}

\item{fixBAinitClarcut}{}

\item{initCLcutRatio}{}

\item{PAR}{}

\item{TAir}{}

\item{VPD}{}

\item{Precip}{}

\item{CO2}{}

\item{P0}{}

\item{initVar}{}

\item{soilC}{}

\item{weatherYasso}{}

\item{litterSize}{}

\item{soilCtot}{}

\item{defaultThin}{}

\item{ClCut}{}

\item{energyCut}{}

\item{inDclct}{}

\item{inAclct}{}

\item{yassoRun}{}

\item{smoothP0}{}

\item{smoothETS}{}

\item{smoothYear}{}

\item{tapioPars}{}

\item{thdPer}{}

\item{limPer}{}

\item{ftTapioPar}{}

\item{tTapioPar}{}

\item{GVrun}{}

\item{thinInt}{}

\item{fertThin}{}

\item{nYearsFert}{}

\item{protect}{}

\item{mortMod}{}

\item{ECMmod}{}

\item{pECMmod}{}

\item{pPRELESgv}{preles parameters for ground vegetation (only runs if layerPRELES==1, other wise 1 parameter set is used for the whole ecosystem)}

\item{layerPRELES}{flag indicating if preles is going to be run by layer with species specific parameters or using 1 parameter set for the whole forest}
}
\description{
function to run the PREBAS model for a single site!
}
