% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebas.r
\name{prebas}
\alias{prebas}
\title{prebas model function}
\usage{
prebas(
  nYears,
  pCROBAS = pCROB,
  pHcMod = pHcM,
  pPRELES = pPREL,
  pYASSO = pYAS,
  pAWEN = parsAWEN,
  etmodel = 0,
  siteInfo = NA,
  thinning = NA,
  initClearcut = initSeedling.def,
  fixBAinitClarcut = 1,
  initCLcutRatio = NA,
  PAR,
  TAir,
  VPD,
  Precip,
  CO2,
  P0 = NA,
  initVar = NA,
  soilC = NA,
  weatherYasso = NA,
  litterSize = litterSizeDef,
  soilCtot = NA,
  defaultThin = 1,
  ClCut = 1,
  energyCut = 0,
  inDclct = NA,
  inAclct = NA,
  yassoRun = 0,
  smoothP0 = 1,
  smoothETS = 1,
  smoothYear = 5,
  tapioPars = pTapio,
  thdPer = 0.5,
  limPer = 0.5,
  ftTapioPar = ftTapio,
  tTapioPar = tTapio,
  GVrun = 1,
  thinInt = -999,
  fertThin = 0,
  nYearsFert = 20,
  protect = 0,
  mortMod = 1,
  ECMmod = 0,
  pECMmod = parsECMmod
)
}
\arguments{
\item{nYears}{number of years of the simulations}

\item{pCROBAS}{}

\item{pHcMod}{}

\item{pPRELES}{}

\item{pYASSO}{}

\item{pAWEN}{}

\item{etmodel}{}

\item{siteInfo}{}

\item{thinning}{}

\item{initClearcut}{}

\item{fixBAinitClarcut}{}

\item{initCLcutRatio}{}

\item{PAR}{}

\item{TAir}{}

\item{VPD}{}

\item{Precip}{}

\item{CO2}{}

\item{P0}{}

\item{initVar}{}

\item{soilC}{}

\item{weatherYasso}{}

\item{litterSize}{}

\item{soilCtot}{}

\item{defaultThin}{}

\item{ClCut}{}

\item{energyCut}{}

\item{inDclct}{}

\item{inAclct}{}

\item{yassoRun}{}

\item{smoothP0}{}

\item{smoothETS}{}

\item{smoothYear}{}

\item{tapioPars}{}

\item{thdPer}{}

\item{limPer}{}

\item{ftTapioPar}{}

\item{tTapioPar}{}

\item{GVrun}{}

\item{thinInt}{}

\item{fertThin}{}

\item{nYearsFert}{}

\item{protect}{}

\item{mortMod}{}

\item{ECMmod}{}

\item{pECMmod}{}
}
\description{
function to run the PREBAS model for a single site!
}
