% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSitePrebas.r
\name{regionPrebas}
\alias{regionPrebas}
\title{Title}
\usage{
regionPrebas(
  multiSiteInit,
  HarvLim = NA,
  minDharv = 999,
  cutAreas = NA,
  clcutlimiter = NA,
  compHarv = 0,
  thinFact = 0.25,
  ageHarvPrior = 0,
  siteOrder = NA,
  fertThin = 0,
  nYearsFert = 20,
  yearFert = NULL,
  deltaSiteTypeFert = 1,
  oldLayer = 0,
  startSimYear = 1
)
}
\arguments{
\item{clcutlimiter}{max share of harvested volume from clear cuts (range 0:1). Remainder towards harvLim needs to be filled by thinnings. Length either 1 or maxYears. Note: internally superseedes cutarea limit for clearcuts.}
}
\description{
Title
}
