% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransectRun.R
\name{TransectRun}
\alias{TransectRun}
\title{Transect runs for pine, spruce, or mixed forests}
\usage{
TransectRun(
  SiteType = NA,
  initVar = NA,
  species = NA,
  nYears = 100,
  pCROBAS = pCROB,
  AgeEffect = F,
  defaultThin = 1,
  ClCut = 1,
  energyCut = 0,
  mortMod = 1
)
}
\arguments{
\item{SiteType}{A value between 1 to 5.}

\item{initVar}{initial state of the forest. Array with nSites,VarsIn,nLayers dimentions: nSites=number of sites (7); VarsIn = initial state input variables (speciesID, age(if NA is calculated by initialAgeSeedl function),h,dbh,ba(by layer),hc,Ac(NA))}

\item{species}{If the initial state of the forest is not provided (initVar) Should be 'Pine', or 'Spruce', or 'Birch' or 'Mixed'}

\item{nYears}{Number of years to run the model.Default value is 100.}

\item{pCROBAS}{(47 x nSpecies matrix) Matrix of parameter sets, each column corresponds to a species. Default values pCROBAS = pCROB are the parameter sets for Scots pine (Pinus sylvestris), Norway spruce (Picea abies), Silver birch (Betula pendula), European beech (Fagus sylvatica), Maritime pine (Pinus pinaster), Blue gum (Eucalyptus globulus), Black locust (Robinia pseudoacacia), Populus(in Romania), Eucalyptus grandis x Eucalyptus urophylla (in Paraguay), and Norway spruce(in Germany). Default is pCROB, print(pCROB) to see the parameter values and names.}

\item{AgeEffect}{Carbon allocation strategy for seedlings. When True, the early growth will decrease due to age effect,by setting pCROBAS[45,]<-0.3.}

\item{defaultThin}{If defaultThin = 1 (default) Finnish standard management practices are applied (ref).}

\item{ClCut}{Vector of Diameter (cm) threshold for clearcut. Each element correspond to a layer of the stand, if only one value is provided the same value is applied to all the layers. The different elements of the vector are for the different layers. The dominant species (highest basal area) is considered for clearcut.}

\item{energyCut}{Energy cutting strategy will be applied if set to 1. Default value is 0.}

\item{mortMod}{flag for the mortality model selection (1= Reineke, 2= random (Siilipehto, 2020), 3= both models)}
}
\value{
The output from multiPrebas()
}
\description{
The simulations of pine or spruce or mixed forests at 7 selected points over a North-South transect. Start from seedlings!
}
\examples{
Pine.1 <- TransectRun(SiteType = 1, species = "Pine")
plot(Pine.1$multiOut[1, , 30, 1, 1], type = "l")

}
