% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransectRun.R
\name{TransectRun}
\alias{TransectRun}
\title{Transect runs for pine, spruce, or mixed forests}
\usage{
TransectRun(
  SiteType = NA,
  initVar = NA,
  species = NA,
  nYears = 100,
  pCROBAS = pCROB,
  pPRELES = pPREL,
  AgeEffect = F,
  defaultThin = 1,
  multiThin = NA,
  multiNthin = NA,
  ClCut = 1,
  energyCut = 0,
  GVrun = 1,
  pHcMod = pHcM,
  etmodel = 0,
  pYASSO = pYAS,
  pAWEN = parsAWEN,
  multiInitClearCut = NA,
  fixBAinitClearcut = 1,
  initCLcutRatio = NA,
  multiP0 = NA,
  soilC = NA,
  weatherInput = NULL,
  weatherYasso = NA,
  litterSize = litterSizeDef,
  soilCtot = NA,
  inDclct = NA,
  inAclct = NA,
  yassoRun = 0,
  smoothP0 = 1,
  smoothETS = 1,
  smoothYear = 5,
  HcModV = HcModV_def,
  tapioPars = pTapio,
  thdPer = NA,
  limPer = NA,
  ftTapioPar = ftTapio,
  tTapioPar = tTapio,
  thinInt = -999,
  mortMod = 1,
  ECMmod = 0,
  layerPRELES = 0,
  LUEtrees = pLUEtrees,
  LUEgv = pLUEgv,
  alpharNcalc = FALSE,
  p0currClim = NA,
  TcurrClim = NA,
  PcurrClim = NA,
  ingrowth = FALSE,
  soilPar = NA,
  siteInfoDist = NA,
  modVersion = "multiSite"
)
}
\arguments{
\item{SiteType}{A value between 1 to 5.}

\item{initVar}{initial state of the forest. Array with nSites,VarsIn,nLayers dimentions: nSites=number of sites (7); VarsIn = initial state input variables (speciesID, age(if NA is calculated by initialAgeSeedl function),h,dbh,ba(by layer),hc,Ac(NA))}

\item{species}{If the initial state of the forest is not provided (initVar) Should be 'Pine', or 'Spruce', or 'Birch' or 'Mixed'}

\item{nYears}{Number of years to run the model.Default value is 100.}

\item{pCROBAS}{(47 x nSpecies matrix) Matrix of parameter sets, each column corresponds to a species. Default values pCROBAS = pCROB are the parameter sets for Scots pine (Pinus sylvestris), Norway spruce (Picea abies), Silver birch (Betula pendula), European beech (Fagus sylvatica), Maritime pine (Pinus pinaster), Blue gum (Eucalyptus globulus), Black locust (Robinia pseudoacacia), Populus(in Romania), Eucalyptus grandis x Eucalyptus urophylla (in Paraguay), and Norway spruce(in Germany). Default is pCROB, print(pCROB) to see the parameter values and names.}

\item{pPRELES}{A vector of PRELES parameter. Default is the boreal forest version pPREL.}

\item{AgeEffect}{Carbon allocation strategy for seedlings. When True, the early growth will decrease due to age effect,by setting pCROBAS[45,]<-0.3.}

\item{defaultThin}{If defaultThin = 1 (default) Finnish standard management practices are applied (ref).}

\item{multiThin}{A array with thinnig inputs. Three dimensions: nSites x maxThin x 9. The first dimension is the number of sites. The second dimension, maxThin, is the number of thinnings. For the third demention, element 1 is year from the start of the simulation; element 2 is siteID; element 3 layer where thinnings are carried out; element 4 to 7 stand variables (H, D, B, Hc); element 8 parameter that indicates if the stand variables (column 4:7) are provided as fraction of the actual model outputs (value=1 means that fraction is used); element 9 is the stand density after thinning if its value is not -999.}

\item{multiNthin}{A matrix with thinning inputs. Rows correspond to a thinning event. Column 1 year from the start of the simulation; column 2 is siteID; column 3 layer where thinnings are carried out; column 4 to 7 stand variables (H, D, B, Hc); column 8 parameter that indicates if the stand variables (column 4:7) are provided as fraction of the actual model outputs (value=1 means that fraction is used); column 9 is the stand density after thinning if its value is not -999; colum 10 is Sapwood area of average tree at crown base (m2) if its value is not -999 (see examples).}

\item{ClCut}{Vector of Diameter (cm) threshold for clearcut. Each element correspond to a layer of the stand, if only one value is provided the same value is applied to all the layers. The different elements of the vector are for the different layers. The dominant species (highest basal area) is considered for clearcut.}

\item{energyCut}{Energy cutting strategy will be applied if set to 1. Default value is 0.}

\item{GVrun}{}

\item{pHcMod}{}

\item{etmodel}{}

\item{pYASSO}{}

\item{pAWEN}{}

\item{multiInitClearCut}{A Matrix: matrix(initClearcut,NoOfSites,5,byrow = T), where initClearcut includes those 5 variables H,dbh,BA,HC,AC, same with 'initSeedling.def'}

\item{fixBAinitClearcut}{}

\item{initCLcutRatio}{}

\item{multiP0}{}

\item{soilC}{}

\item{weatherInput}{##list of weather inputs for PRELES: each variables is a matrix with nrow=nSites,columns=number of days in the simulations}

\item{weatherYasso}{}

\item{litterSize}{}

\item{soilCtot}{}

\item{inDclct}{}

\item{inAclct}{}

\item{yassoRun}{}

\item{smoothP0}{}

\item{smoothETS}{}

\item{smoothYear}{}

\item{HcModV}{flag for the Hc model: 1 use the pipe model defined in the HcPipeMod function, different from 1 uses empirical models; default value (HcModV_def) is 1}

\item{tapioPars}{}

\item{thdPer}{}

\item{limPer}{}

\item{ftTapioPar}{}

\item{tTapioPar}{}

\item{thinInt}{}

\item{mortMod}{flag for the mortality model selection (1= Reineke, 2= random (Siilipehto, 2020), 3= both models)}

\item{ECMmod}{flag for the ECM modelling activation 1 -> model ECM according to Makela et al. 2022, 0 -> no ECM modelling}

\item{layerPRELES}{}

\item{LUEtrees}{light use efficiency parameters for tree species}

\item{LUEgv}{light use efficiency parameter for ground vegetation}

\item{alpharNcalc}{#alphar calculations based on Nitrogen availability. deafault value is FALSE (no nitrogen impact). =1calculates N uptake}

\item{p0currClim}{# vector of average annual P0 for the climIDs at current climate. if NA the first five years of the simulations will be used to calculate it.}

\item{TcurrClim}{# vector of average annual temperature for the climIDs at current climate. if NA the first five years of the simulations will be used to calculate it.}

\item{PcurrClim}{# vector of average annual precipitation for the climIDs current climate. if NA the first five years of the simulations will be used to calculate it.}

\item{ingrowth}{# flag to simulate ingrowth}

\item{soilPar}{# input a matrix (dim=nSites,3 ) with soil depth, FC, WP, for each site if NA uses the default values}

\item{siteInfoDist}{# UNDER DEVELOPMENT! external data for wind risk modelling. Switches on wind risk calculations; output: outDist. Matrix with dims(nSites, 4): 1 = 10a max windspeed, m/s; 2: time since thining (years; for initialisation, but currently used throughout simulations); 3: soiltype (0 = mineral, coarse; 1 = mineral, fine; 2 = organic); 4: shallow soil (0 = F, >30cm; 1 = T, <30cm)}

\item{modVersion}{# model version to use in the simulations it can be multiSite or region}
}
\value{
The output from multiPrebas()
}
\description{
The simulations of pine or spruce or mixed forests at 7 selected points over a North-South transect. Start from seedlings!
}
\examples{
# Qucik examples
Pine.1 <- TransectRun(SiteType = 1, species = "Pine")
plot(Pine.1$multiOut[1, , 30, 1, 1], type = "l")

#' An example of initail sates
InitStands<-matrix(c(1,NA,1.5,0.5,0.0431969,0.,0.),nrow = 7,ncol=7,byrow = T)
Pine.i <-TransectRun(SiteType = 1, initVar = InitStands)
plot(Pine.i$multiOut[1, , 30, 1, 1], type = "l")

#An example of thinning
thinnings<-array(0,c(7,2,10))
thinnings[,,1]<- c(rep(20,7),rep(40,7)) #year from the start of the simulation
thinnings[,,2]<- 1:7 #siteID
thinnings[,,3]<- 1 #layer where thinnings are carried out
thinnings[,,4]<- 1.1 #H
thinnings[,,5]<- 1.1#D
thinnings[,,6]<- 0.5#B
thinnings[,,7]<- 1#Hc
thinnings[,,8]<- 1 #if the stand variables (column 4:7) are provided as fraction of the actual model outputs
thinnings[,,9:10] <- -999 #9 the stand density after thinning if its value is not -999.
Pine.thin <-TransectRun(SiteType = 1, initVar = InitStands,multiThin = thinnings,multiNthin = rep(2,7))
plot(Pine.thin$multiOut[1, , 30, 1, 1], type = "l")

#An example of new settings for seedlings after a clear cut
initClearcut<-c(1.5100000,0.8000000,0.0631969,0.2000000,NA) # H,dbh,BA,HC,AC
afterClearCut <- matrix(initClearcut,7,5,byrow = T)
Pine.newseed <-TransectRun(SiteType = 1, initVar = InitStands,multiThin = thinnings,multiNthin = rep(2,7),multiInitClearCut = afterClearCut)
plot(Pine.newseed$multiOut[1, , 30, 1, 1], type = "l")

}
